Update EnemiesWithItems.cs
Hey, I had taken a look at your code in VisualStudio and with the new R2API and BepInEx updates I noticed that ConfigWrapper and Config.Wrap were obsolete. I changed ConfigWrapper to ConfigEntry, and Config.Wrap to Config.Bind. This switched the default value and description, but I tested it and seems to still work. Wasn't sure if this would help or not, I don't code very much. But I figured I'd say something in case you ended up needing to update it anyways.